The API is expected to use `context` when available.
