+++ client.go
@@ -6,7 +6,6 @@ import (
 	"bytes"
 	"crypto/tls"
 	"encoding/binary"
-	"golang.org/x/net/context"
 	"io"
 	"net"
 	"strings"
@@ -419,16 +418,6 @@ func Dial(network, address string) (conn *Conn, err error) {
 	return conn, nil
 }
 
-// ExchangeContext performs a synchronous UDP query, like Exchange. It
-// additionally obeys deadlines from the passed Context.
-func ExchangeContext(ctx context.Context, m *Msg, a string) (r *Msg, err error) {
-	client := Client{Net: "udp"}
-	r, _, err = client.ExchangeContext(ctx, m, a)
-	// ignorint rtt to leave the original ExchangeContext API unchanged, but
-	// this function will go away
-	return r, err
-}
-
 // ExchangeConn performs a synchronous query. It sends the message m via the connection
 // c and waits for a reply. The connection c is not closed by ExchangeConn.
 // This function is going away, but can easily be mimicked:
@@ -487,20 +476,4 @@ func DialTimeoutWithTLS(network, address string, tlsConfig *tls.Config, timeout 
 		return nil, err
 	}
 	return conn, nil
-}
-
-// ExchangeContext acts like Exchange, but honors the deadline on the provided
-// context, if present. If there is both a context deadline and a configured
-// timeout on the client, the earliest of the two takes effect.
-func (c *Client) ExchangeContext(ctx context.Context, m *Msg, a string) (r *Msg, rtt time.Duration, err error) {
-	var timeout time.Duration
-	if deadline, ok := ctx.Deadline(); !ok {
-		timeout = 0
-	} else {
-		timeout = deadline.Sub(time.Now())
-	}
-	// not passing the context to the underlying calls, as the API does not support
-	// context. For timeouts you should set up Client.Dialer and call Client.Exchange.
-	c.Dialer = &net.Dialer{Timeout: timeout}
-	return c.Exchange(m, a)
 }
